Don't run opensm on compute nodes
With lots of hosts, opensm can require a non-trivial amount of CPU
time. So it shouldn't be run on a compute node where it can disturb
batch jobs. Also, with lots of hosts, if all of them run opensm in
standby mode I guess there will be a lot of polling on the network to
check whether the master is still up.