Remove ansible from compute kickstart_packages
As of CentOS 7.4 ansible was moved from the EPEL repo to the CentOS
extras repo. For some reason it doesn't work to add centos extras to
the repo list in kickstart, hence ansible must be installed separately
in the kickstart %post section. It must also be removed from the list
of packages to install in kickstart, which is what this patch does.